direction change fixed
