+++ Command.c
@@ -1,40 +1,64 @@
 #include "Command.h"
 
+#define NUMBER_OF_CYCLE_PER_DEGREE 6
+uint32_t DEGREE_OF_TURN = 1;
+uint32_t NUMBER_OF_TURN = 0;
+
+
 void FORWARD(){
+	
+	// Turn off interrupt
+	NVIC_DisableIRQ(PWM0_IRQn);
 	
 	// Set Motor
 	MOTOR_DRIVER_IN1_PORT->SET |= MOTOR_DRIVER_IN1_MASK;
 	MOTOR_DRIVER_IN2_PORT->CLR |= MOTOR_DRIVER_IN2_MASK;
 	MOTOR_DRIVER_IN3_PORT->SET |= MOTOR_DRIVER_IN3_MASK;
 	MOTOR_DRIVER_IN4_PORT->CLR |= MOTOR_DRIVER_IN4_MASK;
-	SET_MOTOR_POWER(MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT, MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT);
 	
 	// Set Signal 
 	BACK_SIGNAL_PORT->CLR |= BACK_SIGNAL_MASK;
 	FORWARD_SIGNAL_PORT->SET |= FORWARD_SIGNAL_MASK;
+	Finish_Signal();
 	
 }
 
 void BACK (){
+	
+	// Turn off interrupt
+	NVIC_DisableIRQ(PWM0_IRQn);
+	
 	// Set Motor
 	MOTOR_DRIVER_IN1_PORT->CLR |= MOTOR_DRIVER_IN1_MASK;
 	MOTOR_DRIVER_IN2_PORT->SET |= MOTOR_DRIVER_IN2_MASK;
 	MOTOR_DRIVER_IN3_PORT->CLR |= MOTOR_DRIVER_IN3_MASK;
 	MOTOR_DRIVER_IN4_PORT->SET |= MOTOR_DRIVER_IN4_MASK;
-	SET_MOTOR_POWER(MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT, MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT);
 	
 	// Set Signal 
 	BACK_SIGNAL_PORT->SET |= BACK_SIGNAL_MASK;
 	FORWARD_SIGNAL_PORT->CLR |= FORWARD_SIGNAL_MASK;
+	Finish_Signal();
 }
 
 void STOP(){
-	SET_MOTOR_POWER(0, 0);
+	
+	// Turn off interrupt
+	NVIC_DisableIRQ(PWM0_IRQn);
+	
+	MOTOR_DRIVER_IN1_PORT->SET |= MOTOR_DRIVER_IN1_MASK;
+	MOTOR_DRIVER_IN2_PORT->SET |= MOTOR_DRIVER_IN2_MASK;
+	MOTOR_DRIVER_IN3_PORT->SET |= MOTOR_DRIVER_IN3_MASK;
+	MOTOR_DRIVER_IN4_PORT->SET |= MOTOR_DRIVER_IN4_MASK;
+	
 	BACK_SIGNAL_PORT->CLR |= BACK_SIGNAL_MASK;
 	FORWARD_SIGNAL_PORT->CLR |= FORWARD_SIGNAL_MASK;
+	Finish_Signal();
 }
 
 void RIGHT(uint32_t DEGREE){
+	
+	// Reset
+	NUMBER_OF_TURN = 0;
 	
 	// Set Interrupt for 90 degrees
 	NVIC_EnableIRQ(PWM0_IRQn);
@@ -47,7 +71,6 @@ void RIGHT(uint32_t DEGREE){
 	MOTOR_DRIVER_IN2_PORT->SET |= MOTOR_DRIVER_IN2_MASK;
 	MOTOR_DRIVER_IN3_PORT->SET |= MOTOR_DRIVER_IN3_MASK;
 	MOTOR_DRIVER_IN4_PORT->CLR |= MOTOR_DRIVER_IN4_MASK;
-	SET_MOTOR_POWER(MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT, MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT);
 	
 	// Set Signal 
 	BACK_SIGNAL_PORT->CLR |= BACK_SIGNAL_MASK;
@@ -58,6 +81,9 @@ void RIGHT(uint32_t DEGREE){
 
 void LEFT(uint32_t DEGREE){
 	
+	//Reset
+	NUMBER_OF_TURN = 0;
+	
 	// Set Interrupt for 90 degrees
 	NVIC_EnableIRQ(PWM0_IRQn);
 	NVIC_ClearPendingIRQ(PWM0_IRQn);
@@ -69,7 +95,6 @@ void LEFT(uint32_t DEGREE){
 	MOTOR_DRIVER_IN2_PORT->CLR |= MOTOR_DRIVER_IN2_MASK;
 	MOTOR_DRIVER_IN3_PORT->CLR |= MOTOR_DRIVER_IN3_MASK;
 	MOTOR_DRIVER_IN4_PORT->SET |= MOTOR_DRIVER_IN4_MASK;
-	SET_MOTOR_POWER(MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT, MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT);
 	
 	// Set Signal 
 	BACK_SIGNAL_PORT->CLR |= BACK_SIGNAL_MASK;
@@ -86,4 +111,27 @@ void TESTING (void){
 }
 
 void FINISH(){
+	Request("FINISH\r\n");
 }
+
+void PWM0_IRQHandler(){
+	
+	PWM_MOTOR->IR |= 1;
+	NUMBER_OF_TURN += 4;
+	
+	if(NUMBER_OF_TURN >= NUMBER_OF_CYCLE_PER_DEGREE * DEGREE_OF_TURN){
+		// Turn off interrupt
+		PWM_MOTOR->MCR &= ~1;
+		NVIC_DisableIRQ(PWM0_IRQn);
+		
+		MOTOR_DRIVER_IN1_PORT->SET |= MOTOR_DRIVER_IN1_MASK;
+		MOTOR_DRIVER_IN2_PORT->SET |= MOTOR_DRIVER_IN2_MASK;
+		MOTOR_DRIVER_IN3_PORT->SET |= MOTOR_DRIVER_IN3_MASK;
+		MOTOR_DRIVER_IN4_PORT->SET |= MOTOR_DRIVER_IN4_MASK;
+		
+		Finish_Signal();
+		NUMBER_OF_TURN = 0;
+	}
+}
+
+
