+++ MOTOR.c
@@ -27,12 +27,17 @@ void MOTOR_Init() {
 	
 	PWM_MOTOR->TCR = (1 << 0 | 1 << 3);
 	
-	SET_MOTOR_POWER(0, 0);
+	SET_MOTOR_POWER(MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT, MAX_MOTOR_DUTY_CYCLE * MOTOR_POWER_IN_PERCENT);
 	
 	MOTOR_DRIVER_IN1_PORT->DIR |= MOTOR_DRIVER_IN1_MASK;
 	MOTOR_DRIVER_IN2_PORT->DIR |= MOTOR_DRIVER_IN2_MASK;
 	MOTOR_DRIVER_IN3_PORT->DIR |= MOTOR_DRIVER_IN3_MASK;
 	MOTOR_DRIVER_IN4_PORT->DIR |= MOTOR_DRIVER_IN4_MASK;
+	
+	MOTOR_DRIVER_IN1_PORT->SET |= MOTOR_DRIVER_IN1_MASK;
+	MOTOR_DRIVER_IN2_PORT->SET |= MOTOR_DRIVER_IN2_MASK;
+	MOTOR_DRIVER_IN3_PORT->SET |= MOTOR_DRIVER_IN3_MASK;
+	MOTOR_DRIVER_IN4_PORT->SET |= MOTOR_DRIVER_IN4_MASK;
 	
 }
 
@@ -53,19 +58,3 @@ void SET_MOTOR_POWER(uint32_t left, uint32_t right){
 	PWM_MOTOR->LER |= (3 << 5);
 }
 
-void PWM0_IRQHandler(){
-	
-	PWM_MOTOR->IR |= 1;
-	NUMBER_OF_TURN += 1;
-	
-	if(NUMBER_OF_TURN >= NUMBER_OF_CYCLE_PER_DEGREE * DEGREE_OF_TURN){
-		// Turn off interrupt
-		PWM_MOTOR->MCR &= ~1;
-		NVIC_DisableIRQ(PWM0_IRQn);
-		
-		SET_MOTOR_POWER(0, 0);
-		
-		Finish_Signal();
-		NUMBER_OF_TURN = 0;
-	}
-}
