+++ Objects/MATER.build_log.htm
@@ -45,7 +45,7 @@ compiling TEST.c...
 assembling startup_LPC407x_8x_177x_8x.s...
 compiling system_LPC407x_8x_177x_8x.c...
 linking...
-Program Size: Code=5052 RO-data=260 RW-data=144 ZI-data=1040  
+Program Size: Code=5120 RO-data=260 RW-data=120 ZI-data=1040  
 ".\Objects\MATER.axf" - 0 Error(s), 0 Warning(s).
 
 <h2>Software Packages used:</h2>
@@ -78,7 +78,7 @@ Package Vendor: Keil
       Include file:  Device\Include\system_LPC407x_8x_177x_8x.h
       Source file:   Device\Source\ARM\startup_LPC407x_8x_177x_8x.s
       Source file:   Device\Source\system_LPC407x_8x_177x_8x.c
-Build Time Elapsed:  00:00:05
+Build Time Elapsed:  00:00:04
 </pre>
 </body>
 </html>
