+++ Objects/ultrasonic.d
@@ -16,3 +16,7 @@
 .\objects\ultrasonic.o: PWM.h
 .\objects\ultrasonic.o: Signal.h
 .\objects\ultrasonic.o: Signal.h
+.\objects\ultrasonic.o: HM10.h
+.\objects\ultrasonic.o: Serial.h
+.\objects\ultrasonic.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdio.h
+.\objects\ultrasonic.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\string.h
